+++ src/main/java/org/team6/controller/SettingsPageController.java
@@ -7,8 +7,8 @@ import java.util.ResourceBundle;
 
 import org.team6.model.Notification;
 import org.team6.model.NotificationBackend;
-import org.team6.model.Recommendations.RecommendationObserver;
-import org.team6.model.Recommendations.RecommendationsBackend;
+import org.team6.model.recommendations.RecommendationObserver;
+import org.team6.model.recommendations.RecommendationsBackend;
 import org.team6.view.PageStarter;
 
 import javafx.event.ActionEvent;
@@ -20,16 +20,6 @@ import javafx.scene.control.Spinner;
 import javafx.scene.control.SpinnerValueFactory;
 import javafx.scene.control.ToggleButton;
 import javafx.scene.layout.AnchorPane;
-
-import org.team6.model.Notification;
-import org.team6.model.NotificationBackend;
-import org.team6.model.Recommendations.RecommendationsBackend;
-import org.team6.view.PageStarter;
-
-import java.net.URL;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.ResourceBundle;
 
 public class SettingsPageController implements Initializable, RecommendationObserver {
     @FXML
