+++ src/main/java/org/team6/model/products/Dishwasher.java
@@ -1,11 +1,11 @@
-package org.team6.model.Products;
+package org.team6.model.products;
 
 import org.team6.model.EnergyUsageCategory;
 
-public class WashingMachine extends Product {
-    private double energyConsumption; // Energy consumption of this product in kWh/annum for washing machine
+public class Dishwasher extends Product {
+    private double energyConsumption; // Energy consumption of this product in kWh/annum for dishwasher
 
-    public WashingMachine(String name, ProductCategory productCategory, EnergyUsageCategory energySpendingcategory, double energyConsumption, int price) {
+    public Dishwasher(String name, ProductCategory productCategory, EnergyUsageCategory energySpendingcategory, double energyConsumption, int price) {
         super(name, productCategory, energySpendingcategory, price);
         this.energyConsumption = energyConsumption;
     }    
@@ -18,5 +18,4 @@ public class WashingMachine extends Product {
     public String getProductDescription() {
         throw new UnsupportedOperationException("Not supported yet.");
     }
-    
 }
