+++ src/main/java/org/team6/model/products/Oven.java
@@ -1,4 +1,4 @@
-package org.team6.model.Products;
+package org.team6.model.products;
 
 import org.team6.model.EnergyUsageCategory;
