+++ src/main/java/org/team6/model/products/ProductCategory.java
@@ -1,4 +1,4 @@
-package org.team6.model.Products;
+package org.team6.model.products;
 
 
 public enum ProductCategory {
