+++ src/main/java/org/team6/model/recommendations/RecommendationObserver.java
@@ -1,4 +1,4 @@
-package org.team6.model.Recommendations;
+package org.team6.model.recommendations;
 
 public interface RecommendationObserver {
     void update();
