+++ src/main/java/org/team6/model/recommendations/TipRecommendation.java
@@ -1,4 +1,4 @@
-package org.team6.model.Recommendations;
+package org.team6.model.recommendations;
 
 public class TipRecommendation extends Recommendation {
     public TipRecommendation(String title, String text) {
